+++ src/box/config.py
@@ -75,7 +75,6 @@ class PyProjectParser:
     @property
     def env_var_pyapp(self) -> Dict:
         """Return optional pyapp variables as list (if set), otherwise empty dict."""
-        # TODO: Move this to `tool.box.env-vars.pyapp`
         try:
             return self._pyproject["tool"]["box"]["env-vars"]
         except KeyError:
@@ -125,12 +124,17 @@ class PyProjectParser:
         return self._project["version"]
 
 
-def pyproject_writer(key: str, value: Any) -> None:
+def pyproject_writer(key: str, value: Any, category: str = None) -> None:
     """Modify the existing `pyproject.toml` file using `tomlkit`.
 
     Project specific, the table [tools.box] is used. If the table does not exist,
     it is created. If the key does not exist, it is created. If the key exists,
     it is overwritten.
+
+    :param key: Key to write to.
+    :param value: Value to write to key.
+    :param category: If given, will write to ["tool"]["box"]["category"]["key"]["value"]
+
     """
     pyproject_file = Path("pyproject.toml")
     if not pyproject_file.is_file():
@@ -162,7 +166,22 @@ def pyproject_writer(key: str, value: Any) -> None:
         tool_table.append("box", box_table)
         doc.add("tool", tool_table)
 
-    box_table.update({key: value})
+    if not category:
+        edit_table = box_table
+    else:
+        try:
+            edit_table = doc["tool"]["box"][category]
+        except KeyError:
+            doc.add(tomlkit.nl())
+            tool_table = tomlkit.table(True)
+            category_table = tomlkit.table(True)
+            tool_table.append("box", category_table)
+            category_table.add(tomlkit.nl())
+            edit_table = tomlkit.table()
+            category_table.append(category, edit_table)
+            doc.add("tool", tool_table)
+
+    edit_table.update({key: value})
 
     with open(pyproject_file, "w", newline="\n") as f:
         tomlkit.dump(doc, f)